requirements.txt
pytest==7.4.0
requests==2.31.0
jsonschema==4.19.0
allure-pytest==2.13.2
pytest-html==3.2.0